answers to part 4d added

+++ controllers/blogs.js
@@ -1,18 +1,35 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
+
 const Blog = require('../models/Blog')
+const User = require('../models/User')
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   res.status(200).json(blogs.map(p => p.toJSON()))
 })
 
 blogsRouter.post('/', async (req, res, next) => {
-  const newBlog = new Blog({
-    ...req.body
-  })
+  const body = req.body
+  const token = req.token
+  const decodedToken = jwt.verify(token, process.env.JWTSECRET)
+  if (!token || !decodedToken.id) {
+    res.status(401).send({ error: 'unauthorized access' })
+    return
+  }
+
   try {
+    const user = await User.findById(decodedToken.id)
+    const newBlog = new Blog({
+      author: body.author,
+      title: body.title,
+      url: body.url,
+      user: user._id
+    })
     const blog =  await newBlog.save()
-    res.json(blog.toJSON())
+    user.blogs = user.blogs.concat(blog._id)
+    await user.save()
+    return res.json(blog.toJSON())
   } catch (error) {
     next(error)
   }
@@ -44,7 +61,18 @@ blogsRouter.get('/:id', async (req, res, next) => {
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+  const token = req.token
+  const decodedToken = token && jwt.verify(token, process.env.JWTSECRET)
+  if (!token || !decodedToken.id) {
+    res.status(401).send({ error: 'unauthorized access' })
+    return
+  }
+
   try {
+    const blog = await Blog.findById(req.params.id)
+    if (blog.user.toString() !== decodedToken.id) {
+      return res.status(401).json({ error: 'Unautharized. blog doesnot belong to user' })
+    }
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
   } catch (error) {
